Correção do exercício 1007

+++ 1007/1007.py
@@ -1,7 +1,7 @@
-a = float(input(''))
-b = float(input(''))
-c = float(input(''))
-d = float(input(''))
+a = int(input(''))
+b = int(input(''))
+c = int(input(''))
+d = int(input(''))
 
 diferenca = a * b - c * d
